feat(web/engine): basic store explosion test relocated
